requirements.txt
# Core
typer>=0.9.0
rich>=13.7.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Video
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Data
pandas>=2.1.0
pyarrow>=14.0.0
orjson>=3.9.0

# ML/Tracking
# ML/Tracking
mediapipe>=0.10.9
numpy==1.26.4
# PyTorch & OpenMMLab (Phase 2) - Install via scripts/setup_phase2.py to handle CUDA
# torch==2.1.2+cu118
# torchvision==0.16.2+cu118
# mmcv==2.1.0
# mmpose>=1.3.0
# mmdet>=3.0.0

# Logging
loguru>=0.7.0

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0
gradio>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
